- refactor text2text.py to use ov.Tensor instead of TokenizedINputs - this means we can use the AutoTokenizer from transformers instead of the OpenVINO GenAI tokenizer - GenAI tokenizer was preventing models from loading properly for compatability issues with certain jinja templates.
Instead of mucking around with a broken templating system, I found a way to
leverage HF tokenizers instead of the OpenVINO tokenizers model.

Additionaly, the documentation says there are weird formatting constraints for images that I was able to solve dynamically already, so I want to try that again here once I get o vision implementation working.
THe other thing is that many openvino features are reccomended for convience, but I dont care about that. I want th right tool for the job.